Updates pipeline viz [wip]
