+++ app/views/samples/stage_results.html.erb
@@ -1,9 +1,7 @@
 <div id="PipelineStageResults">
   <%= javascript_tag do %>
-    // react_component('PipelineVizCytoscape', {
-    //    stageResults: JSON.parse('<%= raw escape_json(@results) %>')
-    // }, 'PipelineStageResults');
-    react_component('PipelineVizVis', {
+    react_component('PipelineViz', {
+        admin: <%= current_user.admin %>,
         stageResults: JSON.parse('<%= raw escape_json(@results) %>')
     }, 'PipelineStageResults');
   <% end %>
